refactor: share isAttachmentResponse
