+++ src/types/api/index.ts
@@ -2,3 +2,4 @@ export * from './field'
 export * from './form'
 export * from './encrypt_submission'
 export * from './email_submission'
+export * from './submission'
